% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpcv.R
\name{kpcv}
\alias{kpcv}
\title{Projective k-Means Clustering with Varying Dimension}
\usage{
kpcv(X, K = 2, iter = 496, init = c("kmeans", "random"), print.progress = TRUE)
}
\description{
Projective k-Means Clustering with Varying Dimension
}
