% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsubclust-package.R
\docType{package}
\name{Rsubclust}
\alias{Rsubclust}
\alias{Rsubclust-package}
\title{Bayesian Mixture of Subspaces of Different Dimensions}
\description{
Bayesian model for inferring mixtures of subspaces of different dimensions is presented.
}
